# wms-contract/src/main/resources/openapi/wms-v1.yaml
openapi: 3.1.0
info:
  title: WMS v1 HTTP contract
  version: 0.1.0
  description: |
    公开前缀 `/api/wms/v1`，内部前缀 `/internal/wms/v1`。
    enterpriseId/operator 来自已验证 OIDC token；请求中的企业字段仅核验一致性，不能扩大权限。
    数量 JSON 为十进制字符串。时间 RFC3339 UTC。列表 cursor 分页，默认 50、上限 200。
    写入头必须带 Authorization、Idempotency-Key、X-Request-Id；clientOperationId 若出现必须与 Idempotency-Key 一致。
    Seata TCC Try/Confirm/Cancel 不是 REST 资源，不提供自定义 /prepare 或 /decisions。
    OIDC issuer/client/权限映射由环境变量配置，本文件不含密钥。
    本契约覆盖 S1 主数据与后续切片接口形状；未实现切片不得把 200 当成业务已交付。
servers:
  - url: /
security:
  - oidc: []
...
paths:
...
  /api/wms/v1/warehouses/{warehouseId}/moves:
    post:
      operationId: moveStock
      tags: [inventory]
      summary: 移库
      security:
        - oidc: [stock.move]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveRequest'
      responses:
        '202':
          description: 已持久化受理，不是业务终态成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOperation'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses/{warehouseId}/stock-holds:
    post:
      operationId: createStockHold
      tags: [inventory]
      summary: 库存限制
      security:
        - oidc: [stock.hold]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockHoldRequest'
      responses:
        '202':
          description: 已持久化受理，不是业务终态成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOperation'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses/{warehouseId}/stock-holds/{holdId}/releases:
    post:
      operationId: releaseStockHold
      tags: [inventory]
      summary: 释放库存限制
      security:
        - oidc: [stock.releaseHold]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/HoldId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedReasonRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
  /api/wms/v1/warehouses/{warehouseId}/adjustments:
    post:
      operationId: createAdjustment
      tags: [inventory]
      summary: 创建调整
      security:
        - oidc: [adjustment.create]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustmentCreateRequest'
      responses:
        '201':
          description: 已创建资源
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
    get:
      operationId: listAdjustments
      tags: [inventory]
      summary: 调整列表
      security:
        - oidc: [adjustment.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPage'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses/{warehouseId}/adjustments/{adjustmentId}:
    get:
      operationId: getAdjustment
      tags: [inventory]
      summary: 调整详情
      security:
        - oidc: [adjustment.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/AdjustmentId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses/{warehouseId}/adjustments/{adjustmentId}/approvals:
    post:
      operationId: approveAdjustment
      tags: [inventory]
      summary: 审批调整
      security:
        - oidc: [adjustment.approve]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/AdjustmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustmentApprovalRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses/{warehouseId}/adjustments/{adjustmentId}/applications:
    post:
      operationId: applyAdjustment
      tags: [inventory]
      summary: 应用调整
      security:
        - oidc: [adjustment.apply]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/AdjustmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustmentApplyRequest'
      responses:
        '202':
          description: 已持久化受理，不是业务终态成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOperation'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
  /api/wms/v1/warehouses:
    post:
      operationId: createWarehouse
      tags: [masterdata]
      summary: 创建仓库
      security:
        - oidc: [masterdata.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseCreateRequest'
      responses:
        '201':
          description: 已创建资源
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
    get:
      operationId: listWarehouses
      tags: [masterdata]
      summary: 可访问仓列表
      security:
        - oidc: [masterdata.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPage'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses/{warehouseId}:
    get:
      operationId: getWarehouse
      tags: [masterdata]
      summary: 仓库详情
      security:
        - oidc: [masterdata.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      description: issuer 与 client 由 WMS_OIDC_ISSUER / WMS_OIDC_CLIENT_ID 指向已有提供方，仓库不写死密钥
      openIdConnectUrl: https://issuer.example.invalid/.well-known/openid-configuration
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
      description: 与 clientOperationId 一致；同键同内容返回原结果，同键异内容 409
    XRequestId:
      name: X-Request-Id
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
...
    Cursor:
      description: 不透明游标，绑定原查询条件；更换筛选或投影世代后从第一页重新读取。
      name: cursor
      in: query
      required: false
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    WarehouseId:
      name: warehouseId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
    HoldId:
      name: holdId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
    AdjustmentId:
      name: adjustmentId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
  responses:
...
    Error:
      description: 业务或协议错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Id:
      type: string
      minLength: 1
      maxLength: 64
...
    Version:
      type: integer
      format: int64
      minimum: 0
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, requestId, retryable]
      properties:
        code:
          type: string
          enum:
            - INVALID_QUANTITY
            - INVALID_UNIT
            - REQUIRED_LOT
            - INVALID_EXPIRY
            - UNAUTHENTICATED
            - WAREHOUSE_FORBIDDEN
            - RESOURCE_NOT_FOUND
            - IDEMPOTENCY_PAYLOAD_MISMATCH
            - VERSION_CONFLICT
            - STOCK_INSUFFICIENT
            - STOCK_FROZEN
            - EXPIRED_STOCK
            - SERIAL_OWNERSHIP_CONFLICT
            - STALE_OWNER_EPOCH
            - TCC_TRANSACTION_PENDING
            - INVALID_STATE
            - TCC_CONTEXT_MISMATCH
            - TCC_OWNER_CONFLICT
            - ROUTE_EPOCH_STALE
            - RESOURCE_BUDGET_EXCEEDED
            - DEPENDENCY_UNAVAILABLE
            - EFFECT_IDENTITY_CONFLICT
            - STALE_EXECUTION_ATTEMPT
            - EFFECT_ALREADY_APPLIED
            - RECOVERY_PENDING
            - INVALID_ARGUMENT
            - INVALID_PAGE
            - DATABASE_UNAVAILABLE
            - QUERY_OVERLOADED
            - RATE_LIMITED
            - DUPLICATE_DOCUMENT
            - DUPLICATE_INSPECTION
        message: { type: string }
        requestId: { type: string }
        operationId: { $ref: '#/components/schemas/Id' }
        retryable: { type: boolean }
        details:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              field: { type: string }
              code: { type: string }
    CursorPage:
      type: object
      additionalProperties: false
      required: [items, limit]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ResourceEnvelope' }
        nextCursor: { type: string }
        limit: { type: integer, minimum: 1, maximum: 200 }
        asOf: { type: string, format: date-time }
        lagSeconds: { type: integer, minimum: 0 }
    ResourceEnvelope:
      type: object
      additionalProperties: true
      required: [id, version]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        version: { $ref: '#/components/schemas/Version' }
        status: { type: string }
        actions:
          type: array
          items: { type: string }
          description: 交互提示，写入仍以服务端校验为准
    AcceptedOperation:
      type: object
      additionalProperties: false
      required: [operationId, status, statusUrl]
      properties:
        operationId: { $ref: '#/components/schemas/Id' }
        sourceOperationId: { $ref: '#/components/schemas/Id' }
        commandId: { $ref: '#/components/schemas/Id' }
        fulfillmentId: { $ref: '#/components/schemas/Id' }
        receiptId: { $ref: '#/components/schemas/Id' }
        executionId: { $ref: '#/components/schemas/Id' }
        physicalStatus: { type: string }
        stockSyncStatus: { type: string }
        status: { type: string }
        version: { $ref: '#/components/schemas/Version' }
        statusUrl: { type: string }
        safeToRetry: { type: boolean }
...
    VersionedReasonRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        reason: { type: string }
        expectedVersion: { $ref: '#/components/schemas/Version' }
        clientOperationId: { $ref: '#/components/schemas/Id' }
...
    MoveRequest:
      type: "object"
      additionalProperties: false
      required: ["sourceBalanceId","targetLocationId","qty","reason"]
      properties:
        clientOperationId:
          type: "string"
          maxLength: 64
        sourceBalanceId:
          type: "string"
          maxLength: 64
          minLength: 1
        targetLocationId:
          type: "string"
          maxLength: 64
          minLength: 1
        qty:
          type: "string"
          pattern: "^[0-9]{1,14}([.][0-9]{1,6})?$"
          description: "精确十进制；必须大于0"
        unit:
          type: "string"
          maxLength: 64
        reason:
          type: "string"
          maxLength: 128
          minLength: 1
    StockHoldRequest:
      type: "object"
      additionalProperties: false
      required: ["scope","reason"]
      properties:
        clientOperationId:
          type: "string"
          maxLength: 64
        scope:
          $ref: "#/components/schemas/HoldScope"
        qty:
          type: "string"
          pattern: "^[0-9]{1,14}([.][0-9]{1,6})?$"
          description: "精确十进制；必须大于0"
        reason:
          type: "string"
          maxLength: 128
          minLength: 1
        evidenceRefs:
          type: "array"
          items:
            type: "string"
            minLength: 1
            maxLength: 512
          maxItems: 200
...
    AdjustmentCreateRequest:
      type: "object"
      additionalProperties: false
      required: ["balanceId","deltaQty","reason"]
      properties:
        clientOperationId:
          type: "string"
          maxLength: 64
        balanceId:
          type: "string"
          maxLength: 64
          minLength: 1
        deltaQty:
          type: "string"
          pattern: "^-?[0-9]{1,14}([.][0-9]{1,6})?$"
          description: "精确十进制；可正可负"
        reason:
          type: "string"
          maxLength: 128
          minLength: 1
        countLineId:
          type: "string"
          maxLength: 64
        evidenceRefs:
          type: "array"
          items:
            type: "string"
            minLength: 1
            maxLength: 512
          maxItems: 200
        serialActions:
          type: "array"
          items:
            type: "string"
            minLength: 1
            maxLength: 512
          maxItems: 0
          description: "独立调整不处理序列号身份；非空请求应走盘点流程。"
    AdjustmentApprovalRequest:
      type: "object"
      additionalProperties: false
      required: ["decision","expectedVersion"]
      properties:
        clientOperationId:
          type: "string"
          maxLength: 64
        decision:
          type: "string"
          maxLength: 64
          minLength: 1
        reason:
          type: "string"
          maxLength: 128
        expectedVersion:
          type: "integer"
          minimum: 0
    AdjustmentApplyRequest:
      type: "object"
      additionalProperties: false
      required: ["expectedVersion"]
      properties:
        clientOperationId:
          type: "string"
          maxLength: 64
        expectedVersion:
          type: "integer"
          minimum: 0
...
    WarehouseCreateRequest:
      type: "object"
      additionalProperties: false
      required: ["code","name","timezone"]
      properties:
        clientOperationId:
          type: "string"
          maxLength: 64
        code:
          type: "string"
          maxLength: 32
          minLength: 1
        name:
          type: "string"
          maxLength: 128
          minLength: 1
        timezone:
          type: "string"
          maxLength: 64
          minLength: 1
...
    HoldScope:
      type: "object"
      additionalProperties: false
      required: ["balanceId"]
      properties:
        balanceId:
          type: "string"
          maxLength: 64
          minLength: 1
        qty:
          type: "string"
          pattern: "^[0-9]{1,14}([.][0-9]{1,6})?$"
          description: "精确十进制；必须大于0"
...